+++ test/dockerHover.test.ts
@@ -559,6 +559,9 @@ describe("Dockerfile hover", function () {
                     assertRawHover(content, 10, 14, "value4");
                     assertRawHover(content, 11, 11, "value3");
                     assertRawHover(content, 11, 18, "value4");
+
+                    content = "FROM alpine\n" + instruction + " var" + delimiter + "value\nRUN a \\\nb #c \\\n${var}";
+                    assertRawHover(content, 4, 3, "value");
                 });
 
                 it("referenced variable ${var} no value", function () {
@@ -709,6 +712,9 @@ describe("Dockerfile hover", function () {
                     assertRawHover(content, 10, 12, "value4");
                     assertRawHover(content, 11, 11, "value3");
                     assertRawHover(content, 11, 15, "value4");
+
+                    content = "FROM alpine\n" + instruction + " var" + delimiter + "value\nRUN a \\\nb #c \\\n$var";
+                    assertRawHover(content, 4, 3, "value");
                 });
 
                 it("referenced variable $var no value", function () {